requirements.txt
python-telegram-bot
python-twitter